% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMS-gt.R
\name{GTrms}
\alias{GTrms}
\title{Generalized test for the total absolute variation in a random-effects ANOVA model}
\usage{
GTrms(ng, mus, sse, alpha = NULL, rho0 = NULL, Bmc = 10000)
}
\arguments{
\item{ng}{sample sizes for each factor level}

\item{mus}{sample mean values at each factor level}

\item{sse}{total error sum of squares (sum of within-group sample variances)}

\item{alpha}{desired CI coefficient}

\item{rho0}{the null threshold value to test \eqn{H_0: \rho\ge \rho_0}.}

\item{Bmc}{number of Monte Carlo simulations to approximate the distribution of generalized test statistics}
}
\value{
\describe{
  \item{p.value}{ generalized test p-value }
  \item{GCI}{ computed CI for \eqn{\rho} }
  \item{Qt}{ simulated generalized test statistics }
}
}
\description{
Given an one-way random-effects ANOVA model, \eqn{Y_{ij}=\mu+u_i+\epsilon_{ij}}
with \eqn{u_i\sim N(0,\sigma_b^2), \epsilon_{ij}\sim N(0,\sigma_w^2)}, we
want to make inference on the total absolute variation parameter, \eqn{\rho^2=\mu^2+\sigma_b^2+\sigma_w^2}.
A generalized inference method is taken to conduct significance test and calculate CI for \eqn{\rho}.
}
\examples{
## Simulation example: how to get summary stats from data
A = rep(1:10, 5:14)
Y = 0.5+rnorm(length(A))+rnorm(length(unique(A)))[A]
ng = as.vector(table(A))
sse = sum(tapply(Y, A, var)*(ng-1), na.rm=TRUE)
mus = tapply(Y, A, mean)
ans = GTrms(ng,mus,sse, alpha=0.05,rho0=2.5); ans[1:2]
## Z-test
Zrms(ng,mus,sse, alpha=0.05,rho0=2.5)
## PO comparison study
ng = c(9, 10, 10, 10, 5, 10, 10, 10, 10, 10, 10, 10, 2, 10, 10, 10)
mus = c(-0.026,0.447,0.083,-0.103,-2.587,-0.61,0.04,-0.593, 0.963,0.643,-0.2,-1.337,-4.333,-2.807,0.563,-0.797)
sse = 221.037
set.seed(123)
ans = GTrms(ng,mus,sse, alpha=0.1, rho0=3, Bmc=1e4); ans[1:2]
## Z-test
Zrms(ng,mus,sse, alpha=0.1,rho0=3)
}
\references{
Tsui,K.-W. and Weerahandi,S. (1989). Generalized p-Values in Significance Testing of Hypotheses in the Presence of Nuisance Parameters. Journal of the American Statistical Association 84, 602–607.

Weerahandi,S. (1993). Generalized Confidence Intervals. Journal of the American Statistical Association 88, 899–905.

Bai,Y., Wang,Z., Lystig,T., and Wu,B. (2019) Efficient and powerful equivalency test on combined mean and variance with application to diagnostic device comparison studies. arXiv:1908.07979
}
