% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMS-gt.R
\name{Zrms}
\alias{Zrms}
\title{Large-sample Z-tests for the total absolute variation in a random-effects ANOVA model}
\usage{
Zrms(ng, mus, sse, alpha = 0.1, rho0 = 3, REML = TRUE)
}
\arguments{
\item{ng}{sample sizes for each factor level}

\item{mus}{sample mean values at each factor level}

\item{sse}{total error sum of squares (sum of within-group sample variances)}

\item{alpha}{desired CI coefficient}

\item{rho0}{the null threshold value to test \eqn{H_0: \rho\ge \rho_0}}

\item{REML}{whether we use REML or MLE, default to TRUE}
}
\value{
\describe{
  \item{p.value}{ Z-score and Z-Wald test p-values }
  \item{par0}{ estimated parameters, test statistic and its variance, and CI for \eqn{\rho^2}, under the null hypothesis }
  \item{pars}{ estimated parameters, test statistic and its variance, and CI for \eqn{\rho^2} }
}
}
\description{
Given a random-effects ANOVA model, \eqn{Y_{ij}=\mu+u_i+\epsilon_{ij}}
with \eqn{u_i\sim N(0,\sigma_b^2), \epsilon_{ij}\sim N(0,\sigma_w^2)}, we
want to make inference on the total absolute variation parameter, \eqn{\rho^2=\mu^2+\sigma_b^2+\sigma_w^2}.
Large-sample Z-tests can be conducted based on approximating \eqn{\sum_{i,j}Y_{ij}^2} with a normal distribution.
}
\examples{
A = rep(1:10, 5:14)
Y = 0.5+rnorm(length(A))+rnorm(length(unique(A)))[A]
ng = as.vector(table(A))
sse = sum(tapply(Y, A, var)*(ng-1), na.rm=TRUE)
mus = tapply(Y, A, mean)
Zrms(ng,mus,sse, rho0=2.5)
}
\references{
Ndikintum, N. K. and Rao, M. (2016). A Special Inference Problem in Repeated Measures Design Test of Statistical Hypothesis on Accuracy Root Mean Square Application to Pulse Oximetry Studies. Statistics in Biopharmaceutical Research 8, 60–76.

Pennello, G. A. (2002). Sample size for paired repeated measures designs of method compari- son studies: Application to pulse oximetry. ASA Proceedings of the Joint Statistical Meetings pages 2671–2675.

Pennello, G. A. (2003). Comparing monitoring devices when a gold standard in unavailable: Application to pulse oximeters. ASA Proceedings of the Joint Statistical Meetings pages 3256–3263.

Bai,Y., Wang,Z., Lystig,T., and Wu,B. (2019) Efficient and powerful equivalency test on combined mean and variance with application to diagnostic device comparison studies. arXiv:1908.07979
}
